requirements.txt
google-cloud-bigquery==3.22.0
openai==1.25.2
google-cloud-secret-manager==2.20.0

